Updated ConsoleApplication3.cpp
The checks for the secret endings were not defined, so I added them. Additionally, 'Sleep(____)' was causing an error, so I replaced it with a different one with the same function. I would test it for you, but my luck is not good enough to hit that low of a chance.